+++ src/apps/good-ghosting/good-ghosting.definition.ts
@@ -15,6 +15,7 @@ export const GOOD_GHOSTING_DEFINITION = appDefinition({
       id: 'game',
       type: GroupType.POSITION,
       label: 'Games',
+      groupLabel: 'Farms',
     },
   },