+++ src/apps/olympus/ethereum/olympus.s-ohm.token-fetcher.ts
@@ -30,6 +30,7 @@ export class EthereumOlympusSOhmTokenFetcher implements PositionFetcher<AppToken
       appId,
       groupId,
       network,
+      exchangeable: true,
       resolveContract: ({ address, network }) => this.contractFactory.olympusSOhmToken({ address, network }),
       resolveVaultAddresses: () => ['0x04906695d6d12cf5459975d7c3c03356e4ccd460'], // sOHM
       resolveUnderlyingTokenAddress: () => '0x64aa3364f17a4d01c6f1751fd97c2bd3d7e7f1d5', // OHM