+++ src/apps/pika-protocol/optimism/pika-protocol.vault.contract-position-fetcher.ts
@@ -85,6 +85,7 @@ export class OptimismPikaProtocolVaultContractPositionFetcher implements Positio
             label,
             secondaryLabel,
             images,
+            statsItems: [{ label: 'Liquidity', value: buildDollarDisplayItem(totalValueLocked) }],
           },
         };